+++ packages/nft-checkout/src/const/abi/swap-v2.ts
@@ -1,4 +1,4 @@
-export const SWAP_V2 = [
+export const SWAP_V2_ABI = [
   {
     inputs: [
       {