+++ packages/nft-checkout/src/const/bundle.ts
@@ -1,0 +1,1 @@
+export const BUNDLE_SALT_BYTES = 32