+++ packages/nft-checkout/src/types/config.ts
@@ -1,0 +1,5 @@
+export type Config = {
+  UNISWAP_V3_TOKEN_LIST_URL: string
+  TRADER_JOE_TOKEN_LIST_URL: string
+  PANCAKE_SWAP_TOKEN_LIST_URL: string
+}