+++ packages/provider/src/providers/index.ts
@@ -1,5 +1,6 @@
 export * from './metamask'
 export * from './phantom'
 export * from './coinbase'
-export * from './solfare'
+export * from './solflare'
 export * from './near'
+export * from './solflare'